:art: :truck: migrate fixtures to their own modules in test/util
