+++ app/tests/util/credentials.py
@@ -1,0 +1,202 @@
+import asyncio
+from typing import List
+
+import pytest
+
+from app.event_handling.sse_listener import SseListener
+from app.models.tenants import CreateTenantResponse
+from app.routes.issuer import router
+from app.tests.util.ecosystem_connections import FaberAliceConnect, MeldCoAliceConnect
+from shared import RichAsyncClient
+from shared.models.credential_exchange import CredentialExchange
+
+CREDENTIALS_BASE_PATH = router.prefix
+
+
+@pytest.fixture(scope="function")
+async def issue_credential_to_alice(
+    faber_client: RichAsyncClient,
+    credential_definition_id: str,  # pylint: disable=redefined-outer-name
+    faber_and_alice_connection: FaberAliceConnect,
+    alice_member_client: RichAsyncClient,
+    alice_tenant: CreateTenantResponse,
+) -> CredentialExchange:
+    credential = {
+        "protocol_version": "v1",
+        "connection_id": faber_and_alice_connection.faber_connection_id,
+        "indy_credential_detail": {
+            "credential_definition_id": credential_definition_id,
+            "attributes": {"speed": "10"},
+        },
+    }
+
+    alice_credentials_listener = SseListener(
+        topic="credentials", wallet_id=alice_tenant.wallet_id
+    )
+
+    # create and send credential offer- issuer
+    await faber_client.post(
+        CREDENTIALS_BASE_PATH,
+        json=credential,
+    )
+
+    payload = await alice_credentials_listener.wait_for_event(
+        field="connection_id",
+        field_id=faber_and_alice_connection.alice_connection_id,
+        desired_state="offer-received",
+    )
+
+    alice_credential_id = payload["credential_id"]
+
+    # send credential request - holder
+    response = await alice_member_client.post(
+        f"{CREDENTIALS_BASE_PATH}/{alice_credential_id}/request", json={}
+    )
+
+    await alice_credentials_listener.wait_for_event(
+        field="credential_id", field_id=alice_credential_id, desired_state="done"
+    )
+
+    return response.json()
+
+
+@pytest.fixture(scope="function")
+async def meld_co_issue_credential_to_alice(
+    meld_co_client: RichAsyncClient,
+    meld_co_credential_definition_id: str,  # pylint: disable=redefined-outer-name
+    meld_co_and_alice_connection: MeldCoAliceConnect,
+    alice_member_client: RichAsyncClient,
+    alice_tenant: CreateTenantResponse,
+) -> CredentialExchange:
+    credential = {
+        "protocol_version": "v1",
+        "connection_id": meld_co_and_alice_connection.meld_co_connection_id,
+        "indy_credential_detail": {
+            "credential_definition_id": meld_co_credential_definition_id,
+            "attributes": {"speed": "10"},
+        },
+    }
+
+    alice_credentials_listener = SseListener(
+        topic="credentials", wallet_id=alice_tenant.wallet_id
+    )
+
+    # create and send credential offer- issuer
+    await meld_co_client.post(
+        CREDENTIALS_BASE_PATH,
+        json=credential,
+    )
+
+    payload = await alice_credentials_listener.wait_for_event(
+        field="connection_id",
+        field_id=meld_co_and_alice_connection.alice_connection_id,
+        desired_state="offer-received",
+    )
+
+    alice_credential_id = payload["credential_id"]
+
+    # send credential request - holder
+    response = await alice_member_client.post(
+        f"{CREDENTIALS_BASE_PATH}/{alice_credential_id}/request", json={}
+    )
+
+    await alice_credentials_listener.wait_for_event(
+        field="credential_id", field_id=alice_credential_id, desired_state="done"
+    )
+
+    return response.json()
+
+
+@pytest.fixture(scope="function")
+async def issue_alice_creds_and_revoke_unpublished(
+    faber_client: RichAsyncClient,
+    alice_member_client: RichAsyncClient,
+    alice_tenant: CreateTenantResponse,
+    credential_definition_id_revocable: str,  # pylint: disable=redefined-outer-name
+    faber_and_alice_connection: FaberAliceConnect,
+) -> List[CredentialExchange]:
+    faber_conn_id = faber_and_alice_connection.faber_connection_id
+
+    for i in range(3):
+        credential = {
+            "protocol_version": "v1",
+            "connection_id": faber_conn_id,
+            "save_exchange_record": True,
+            "indy_credential_detail": {
+                "credential_definition_id": credential_definition_id_revocable,
+                "attributes": {"speed": str(i)},
+            },
+        }
+
+        await faber_client.post(
+            CREDENTIALS_BASE_PATH,
+            json=credential,
+        )
+
+    num_tries = 0
+    num_credentials_returned = 0
+    while num_credentials_returned != 3 and num_tries < 10:
+        await asyncio.sleep(0.25)
+        alice_cred_ex_response = (
+            await alice_member_client.get(
+                f"{CREDENTIALS_BASE_PATH}?connection_id={faber_and_alice_connection.alice_connection_id}"
+            )
+        ).json()
+        num_credentials_returned = len(alice_cred_ex_response)
+        num_tries += 1
+
+    if num_credentials_returned != 3:
+        raise Exception(
+            f"Expected 3 credentials to be issued; only got {num_credentials_returned}"
+        )
+
+    listener = SseListener(topic="credentials", wallet_id=alice_tenant.wallet_id)
+
+    for cred in alice_cred_ex_response:
+        await alice_member_client.post(
+            f"{CREDENTIALS_BASE_PATH}/{cred['credential_id']}/request", json={}
+        )
+        # add sse listener to wait for credential state "done" for each credential
+        await listener.wait_for_event(
+            field="credential_id", field_id=cred["credential_id"], desired_state="done"
+        )
+
+    cred_ex_response = (
+        await faber_client.get(
+            CREDENTIALS_BASE_PATH + "?connection_id=" + faber_conn_id
+        )
+    ).json()
+
+    assert len(cred_ex_response) == 3
+
+    # revoke all credentials in list
+    for cred in cred_ex_response:
+        await faber_client.post(
+            f"{CREDENTIALS_BASE_PATH}/revoke",
+            json={
+                "credential_definition_id": credential_definition_id_revocable,
+                "credential_exchange_id": cred["credential_id"][3:],
+            },
+        )
+
+    credential_exchange_records = [
+        CredentialExchange(**cred) for cred in cred_ex_response
+    ]
+    return credential_exchange_records
+
+
+@pytest.fixture(scope="function")
+async def issue_alice_creds_and_revoke_published(
+    faber_client: RichAsyncClient,
+    issue_alice_creds_and_revoke_unpublished: List[  # pylint: disable=redefined-outer-name
+        CredentialExchange
+    ],
+) -> List[CredentialExchange]:
+    credential_exchange_records = issue_alice_creds_and_revoke_unpublished
+    # Publish revoked credentials
+    await faber_client.post(
+        f"{CREDENTIALS_BASE_PATH}/publish-revocations",
+        json={"revocation_registry_credential_map": {}},
+    )
+
+    return credential_exchange_records
